\name{aiPlot}
\alias{aiPlot}
\alias{aiPlot.rmx}
\title{Plot Absolute Information of Estimators}
\description{
  The function \code{aiPlot} plots the absolute information (AI) of estimators.
}
\usage{
aiPlot(x, \dots)

\method{aiPlot}{rmx}(x, range.alpha = 1e-6, range.n = 501, 
                    param.digits = 2, ggplot.xlab = "x", 
                    ggplot.ylab = expression(paste(abs(IF(x))^2)),
                    ggplot.ggtitle = NULL,
                    point.col = "#0072B5", point.alpha = 0.4, \dots)
}
\arguments{
  \item{x}{ object of S3 class \code{rmx}. }
  \item{range.alpha}{ alpha-quantile used to compute plot range, which is from
                      \code{range.alpha}-quantile to 1-\code{range.alpha}-quantile 
                      of the fitted model.}
  \item{range.n}{ number of points in the alpha-quantile range, generated by 
                  function \code{seq}. }
  \item{param.digits}{ number of digits used for the estimated parameter values, if
                       default \code{ggplot.ggtitle} is used. }
  \item{ggplot.xlab}{ label(s) of x-axis, recycled if length is equal to 1 and more
                      than 1 parameter has been estimated. }
  \item{ggplot.ylab}{ label(s) of y-axis, recycled if length is equal to 1 and more
                      than 1 parameter has been estimated. }
  \item{ggplot.ggtitle}{ if \code{NULL} a default plot title is. If it is specified,
                         it should have length 1. }
  \item{point.col}{ single color used for colouring the data points. }
  \item{point.alpha}{ alpha used for color shading.}
  \item{\dots}{ further arguments passed through. }
}
\details{
  The function is inspired by the \code{plot}-methods and function \code{PlotIC} 
  of the RobASt-family of packages.
  
  In case of optimally-robust RMX estimators computed with function \code{rmx}
  (S3 class \code{rmx}), the absolute information is plotted which corresponds 
  to the squared length of the influence function. 
}
\value{
  Invisible object of class \code{ggplot}.
}
\references{
  Kohl, M. (2005) \emph{Numerical Contributions to the Asymptotic Theory of Robustness}. 
  Bayreuth: Dissertation.

  Rieder, H. (1994) \emph{Robust Asymptotic Statistics}. New York: Springer.

  Rieder, H., Kohl, M. and Ruckdeschel, P. (2008) The Costs of not Knowing
  the Radius. Statistical Methods and Applications \emph{17}(1) 13-40.
  Extended version: \url{http://r-kurs.de/RRlong.pdf}
  
  M. Kohl, P. Ruckdeschel, and H. Rieder (2010). Infinitesimally Robust Estimation 
  in General Smoothly Parametrized Models. \emph{Statistical Methods and Application}, 
  \bold{19}(3):333-354. 
}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{rmx}}, \code{\link{optIF}}}
\examples{
ind <- rbinom(100, size=1, prob=0.05) 
x <- rnorm(100, mean=ind*3, sd=(1-ind) + ind*9)
res <- rmx(x, eps.lower = 0.01, eps.upper = 0.1)
aiPlot(res)

## plot-method
plot(res, which = 5)

## setting and passing argument
aiPlot(res, ggplot.xlab = "data", ggplot.ylab = "absolute information")
plot(res, which = 5, 
     control = list(aiPlot = list(ggplot.xlab = "data", 
                                  ggplot.ylab = "influence")))
}
\keyword{robust}
