\name{plotIF}
\alias{plotIF}
\title{Plot IF of Optimally Robust RMX Estimator}
\description{
  The function \code{plotIF} plots the influence function (IF) of 
  an optimally robust RMX estimator (S3 class \code{rmx}) computed
  with function \code{rmx}. In addition, the data that was used 
  to compute the estimator is added to the plot, where the point size
  is proportional to the absolute/total information (= length of the IF) 
  attributed to each point.
}
\usage{
plotIF(x, range.alpha = 1e-6, range.n = 501, 
       info.digits = 2, param.digits = 2, 
       ggplot.xlab = "x", ggplot.ylab = "IF(x)",
       ggplot.ggtitle = NULL,
       point.col = "#0072B5", point.alpha = 0.4, 
       point.length.out = 5, point.range = c(1,7))
}
\arguments{
  \item{x}{ object of S3 class \code{rmx}. }
  \item{range.alpha}{ alpha-quantile used to compute plot range, which is from
                      \code{range.alpha}-quantile to 1-\code{range.alpha}-quantile 
                      of the fitted model.}
  \item{range.n}{ number of points in the alpha-quantile range, generated by 
                  function \code{seq}. }
  \item{info.digits}{ number of digits shown for \code{info}, which is the absolut/total
                      information assigned to an observation (= length of the IF). }
  \item{param.digits}{ number of digits used for the estimated parameter values, if
                       default \code{ggplot.ggtitle} is used. }
  \item{ggplot.xlab}{ label(s) of x-axis, recycled if length is equal to 1 and more
                      than 1 parameter has been estimated. }
  \item{ggplot.ylab}{ label(s) of y-axis, recycled if length is equal to 1 and more
                      than 1 parameter has been estimated. }
  \item{ggplot.ggtitle}{ if \code{NULL} default plot titles for each component of the
                         IF are generated. If it is specified, it should have the 
                         same length as the number of parameters. If length is equal 
                         to 1 and more than 1 parameter has been estimated, it will
                         be recycled. }
  \item{point.col}{ single color used for colouring the data points. }
  \item{point.alpha}{ alpha used for color shading.}
  \item{point.length.out}{ number of point sizes used to visualize the absolut/total
                           information assigned to a data point; see argument
                           \code{breaks} of function \code{\link{scale_size}}.}
  \item{point.range}{ a numeric vector of length 2 specifing the minimum and maximum 
                      size of the data points; see argument
                      \code{range} of function \code{\link{scale_size}}}
}
\details{
  The function is similar to the \code{plot}-methods and function \code{PlotIC} 
  of the RobASt-family of packages.
}
\value{
  Invisible object of class \code{ggplot}.
}
\references{
  Kohl, M. (2005) \emph{Numerical Contributions to the Asymptotic Theory of Robustness}. 
  Bayreuth: Dissertation.

  Rieder, H. (1994) \emph{Robust Asymptotic Statistics}. New York: Springer.

  Rieder, H., Kohl, M. and Ruckdeschel, P. (2008) The Costs of not Knowing
  the Radius. Statistical Methods and Applications \emph{17}(1) 13-40.
  Extended version: \url{http://r-kurs.de/RRlong.pdf}
  
  M. Kohl, P. Ruckdeschel, and H. Rieder (2010). Infinitesimally Robust Estimation 
  in General Smoothly Parametrized Models. \emph{Statistical Methods and Application}, 
  \bold{19}(3):333-354. 
}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{rmx}}, \code{\link{optIF}}}
\examples{
ind <- rbinom(100, size=1, prob=0.05) 
x <- rnorm(100, mean=ind*3, sd=(1-ind) + ind*9)
res <- rmx(x, eps.lower = 0.01, eps.upper = 0.1)
plotIF(res)

## plot-method
plot(res, which = 1)

## setting and passing argument
plotIF(res, ggplot.xlab = "data", ggplot.ylab = "influence",
       ggplot.ggtitle = c("Location part of RMX-IF",
                          "Scale part of RMX-IF"))
plot(res, which = 1, 
     control = list(plotIF = list(ggplot.xlab = "data", 
                                  ggplot.ylab = "influence",
                                  ggplot.ggtitle = c("Location part of RMX-IF",
                                                     "Scale part of RMX-IF"))))
}
\keyword{robust}
