\name{riPlot}
\alias{riPlot}
\alias{riPlot.rmx}
\title{Plot Relative Information of Estimators}
\description{
  The function \code{riPlot} plots the relative information (RI) of estimators.
  It is only usefull and implemented for models where at least two parameters
  have to be estimated.
}
\usage{
riPlot(x, \dots)

\method{riPlot}{rmx}(x, range.alpha = 1e-6, range.n = 501, 
        param.digits = 2, ggplot.xlab = "x", 
        ggplot.ylab = NULL, 
        ggplot.ggtitle = NULL,
        point.col = "#0072B5", point.alpha = 0.4, 
        point.length.out = 5, point.range = c(1,7), 
        plot = TRUE, \dots)
}
\arguments{
  \item{x}{ object of S3 class \code{rmx}. }
  \item{range.alpha}{ alpha-quantile used to compute plot range, which is from
                      \code{range.alpha}-quantile to 1-\code{range.alpha}-quantile 
                      of the fitted model.}
  \item{range.n}{ number of points in the alpha-quantile range, generated by 
                  function \code{seq}. }
  \item{param.digits}{ number of digits used for the estimated parameter values, if
                       default \code{ggplot.ggtitle} is used. }
  \item{ggplot.xlab}{ labels of x-axes, recycled if length is equal to 1. }
  \item{ggplot.ylab}{ if \code{NULL} default labels of y-axes are generated. If
                      it is specified, it should have the same length as the number 
                      of parameters. If length is equal to 1, it will be recycled. }
  \item{ggplot.ggtitle}{ if \code{NULL} default plot titles for each component of the
                         IF are generated. If it is specified, it should have the 
                         same length as the number of parameters. If length is equal 
                         to 1, it will be recycled. }
  \item{point.col}{ single color used for colouring the data points. }
  \item{point.alpha}{ alpha used for color shading.}
  \item{point.length.out}{ number of point sizes used to visualize the absolut/total
                           information assigned to a data point; see argument
                           \code{breaks} of function \code{\link{scale_size}}.}
  \item{point.range}{ a numeric vector of length 2 specifing the minimum and maximum 
                      size of the data points; see argument
                      \code{range} of function \code{\link{scale_size}}}
  \item{plot}{ if \code{FALSE}, the plotting is suppressed. Can be used to modify
               the invisibly returned plotting object. }
  \item{\dots}{ further arguments passed through. }
}
\details{
  The function is inspired by the \code{plot}-methods and function \code{PlotIC} 
  of the RobASt-family of packages.
  
  In case of optimally-robust RMX estimators computed with function \code{rmx}
  (S3 class \code{rmx}), the relative information is plotted which corresponds 
  to the absolute value of the the respective component/coordinate of the 
  influence function devided the absolute information. The size of the data 
  points represents the absolute information.
}
\value{
  Invisible object of class \code{ggplot} or a list of \code{ggplot} objects.
}
\references{
  Kohl, M. (2005) \emph{Numerical Contributions to the Asymptotic Theory of Robustness}. 
  Bayreuth: Dissertation.

  Rieder, H. (1994) \emph{Robust Asymptotic Statistics}. New York: Springer.

  Rieder, H., Kohl, M. and Ruckdeschel, P. (2008) The Costs of not Knowing
  the Radius. Statistical Methods and Applications \emph{17}(1) 13-40.
  Extended version: \url{http://r-kurs.de/RRlong.pdf}
  
  M. Kohl, P. Ruckdeschel, and H. Rieder (2010). Infinitesimally Robust Estimation 
  in General Smoothly Parametrized Models. \emph{Statistical Methods and Application}, 
  \bold{19}(3):333-354. 
}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link{rmx}}, \code{\link{optIF}}}
\examples{
ind <- rbinom(100, size=1, prob=0.05) 
x <- rnorm(100, mean=ind*3, sd=(1-ind) + ind*9)
res <- rmx(x, eps.lower = 0.01, eps.upper = 0.1)
riPlot(res)

## plot-method
plot(res, which = 6)

## setting and passing argument
riPlot(res, ggplot.xlab = "data", ggplot.ylab = "relative information",
       ggplot.ggtitle = c("Rel. information of location part",
                          "Rel. information of scale part"))
plot(res, which = 6, 
     control = list(riPlot = list(ggplot.xlab = "data", 
                                  ggplot.ylab = "relative information",
                                  ggplot.ggtitle = c("Rel. information of location part",
                                                     "Rel. information of scale part"))))
}
\keyword{robust}
