\name{rutherford}
\alias{rutherford}
\docType{data}
\title{ Dataset Generated by Rutherford and Geiger (1910) }
\description{
  This dataset was generated by Rutherford and Geiger (1910).
}
\usage{data(rutherford)}
\format{
  A numeric vector including the radioactive decay counts of polonium recorded 
  by Rutherford and Geiger (1910). 2608 time intervals with a length of 7.5 seconds 
  (1/8 minute) were observed and the decays counted in each interval. A total 
  of 10097 decays occurred during the entire observation time.
}
\details{
  The dataset consists of 2608 observations.
}
\source{
  The data is taken from page 701 of Rutherford et al. (1910).
}
\references{
  E. Rutherford, H. Geiger, H. Bateman (1910). The probability variations in the 
  distribution of alpha particles. The London, Edinburgh, and Dublin Philosophical 
  Magazine and Journal of Science, \bold{20}(118):698-707. \doi{10.1080/14786441008636955}.
}
\examples{
  data(rutherford)
  table(rutherford)/length(rutherford)
}
\keyword{datasets}
